thumbnail_system/requirements.txt
numpy
Pillow
openai
pytesseract
requests
python-dotenv